requirements.txt
absl-py==2.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
numpy==1.24.4
openrouteservice==2.3.3
pandas==2.0.3
protobuf==4.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
